Make duplicatedInstances able to ignore attributes and not only replace - calling it from BATCH Update
